LPD-44140 Refactor FDS URL resolvers into the right package
# breaking

## What modules/apps/frontend-data-set/frontend-data-set-api/bnd.bnd

package com.liferay.frontend.data.set.resolver is no longer exported in this bundle

## Why

No longer needed as com.liferay.frontend.data.set.url supersedes the removed package

----

# breaking

## What modules/apps/frontend-data-set/frontend-data-set-api/src/main/java/com/liferay/frontend/data/set/resolver/FDSAPIURLResolver.java

FDSAPIURLResolver has been moved to the com.liferay.frontend.data.set.url package. Any implementor must change the package name of this class

As a side note, FDSAPIURLResolverRegistry interface has also been moved. Nevertheless, no additional implementations are expected for this class other than the one provided by DXP so this does not represent a breaking change for customers.

## Why

This is the result of the refactor that unifies FDS API URL computation. At the moment of this writing, no external developers should be using this API as it's still internal

----
