+++ modules/apps/commerce/commerce-order-content-web/src/main/java/com/liferay/commerce/order/content/web/internal/frontend/data/set/url/PendingCommerceOrderFDSAPIURLResolver.java
@@ -3,15 +3,15 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.commerce.order.content.web.internal.frontend.data.set.resolver;
+package com.liferay.commerce.order.content.web.internal.frontend.data.set.url;
 
 import com.liferay.account.model.AccountEntry;
 import com.liferay.commerce.constants.CommerceWebKeys;
 import com.liferay.commerce.context.CommerceContext;
+import com.liferay.commerce.model.CommerceOrder;
 import com.liferay.commerce.product.model.CommerceChannel;
 import com.liferay.commerce.product.service.CommerceChannelLocalService;
-import com.liferay.frontend.data.set.resolver.FDSAPIURLResolver;
-import com.liferay.petra.string.StringPool;
+import com.liferay.frontend.data.set.url.FDSAPIURLResolver;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.util.StringUtil;
 
@@ -21,17 +21,18 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Balazs Breier
+ * @author Gianmarco Brunialti Masera
  */
 @Component(
-	property = "fds.rest.application.key=/headless-commerce-delivery-order/v1.0/PlacedOrder",
+	property = "fds.rest.application.key=/headless-commerce-delivery-cart/v1.0/Cart",
 	service = FDSAPIURLResolver.class
 )
-public class PlacedCommerceOrderFDSAPIURLResolver implements FDSAPIURLResolver {
+public class PendingCommerceOrderFDSAPIURLResolver
+	implements FDSAPIURLResolver {
 
 	@Override
 	public String getSchema() {
-		return "PlacedOrder";
+		return "Cart";
 	}
 
 	@Override
@@ -42,39 +43,28 @@ public class PlacedCommerceOrderFDSAPIURLResolver implements FDSAPIURLResolver {
 			(CommerceContext)httpServletRequest.getAttribute(
 				CommerceWebKeys.COMMERCE_CONTEXT);
 
-		_accountEntry = commerceContext.getAccountEntry();
-
-		if (_accountEntry == null) {
-			return StringPool.BLANK;
-		}
-
-		_commerceChannel = _commerceChannelLocalService.getCommerceChannel(
-			commerceContext.getCommerceChannelId());
-
-		String externalReferenceCode = StringPool.BLANK;
-
-		if (baseURL.startsWith("/v1.0/channels/by-externalReferenceCode")) {
-			externalReferenceCode = _commerceChannel.getExternalReferenceCode();
-		}
+		AccountEntry accountEntry = commerceContext.getAccountEntry();
+		CommerceChannel commerceChannel =
+			_commerceChannelLocalService.getCommerceChannel(
+				commerceContext.getCommerceChannelId());
+		CommerceOrder commerceOrder = commerceContext.getCommerceOrder();
 
 		return StringUtil.replace(
 			baseURL,
 			new String[] {
-				"{accountExternalReferenceCode}", "{accountId}",
+				"{accountExternalReferenceCode}", "{accountId}", "{cartId}",
 				"{channelExternalReferenceCode}", "{channelId}",
 				"{externalReferenceCode}"
 			},
 			new String[] {
-				_accountEntry.getExternalReferenceCode(),
-				String.valueOf(_accountEntry.getAccountEntryId()),
-				_commerceChannel.getExternalReferenceCode(),
-				String.valueOf(_commerceChannel.getCommerceChannelId()),
-				externalReferenceCode
+				accountEntry.getExternalReferenceCode(),
+				String.valueOf(accountEntry.getAccountEntryId()),
+				String.valueOf(commerceOrder.getCommerceOrderId()),
+				commerceChannel.getExternalReferenceCode(),
+				String.valueOf(commerceContext.getCommerceChannelId()),
+				commerceOrder.getExternalReferenceCode()
 			});
 	}
-
-	private AccountEntry _accountEntry;
-	private CommerceChannel _commerceChannel;
 
 	@Reference
 	private CommerceChannelLocalService _commerceChannelLocalService;
