+++ modules/apps/commerce/commerce-order-content-web/src/test/java/com/liferay/commerce/order/content/web/internal/frontend/data/set/url/PlacedCommerceOrderFDSAPIURLResolverTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.commerce.order.content.web.internal.frontend.data.set.resolver;
+package com.liferay.commerce.order.content.web.internal.frontend.data.set.url;
 
 import com.liferay.account.model.AccountEntry;
 import com.liferay.commerce.constants.CommerceWebKeys;
