+++ modules/apps/frontend-data-set/frontend-data-set-api/bnd.bnd
@@ -9,7 +9,6 @@ Export-Package:\
 	com.liferay.frontend.data.set.model,\
 	com.liferay.frontend.data.set.provider,\
 	com.liferay.frontend.data.set.provider.search,\
-	com.liferay.frontend.data.set.resolver,\
 	com.liferay.frontend.data.set.sort,\
 	com.liferay.frontend.data.set.url,\
 	com.liferay.frontend.data.set.view,\
