--- modules/apps/frontend-data-set/frontend-data-set-api/src/main/resources/com/liferay/frontend/data/set/resolver/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
